Created body that moves to the right using refactored code

+++ include/physics/gravity.h
@@ -1,10 +1,10 @@
 #ifndef GRAVITY_H
 #define GRAVITY_H
 
-#include "body.h"
-#include "vec2.h"
+#include "core/body.h"
+#include "math/vec2.h"
 
-double gravitational_force_magnitude(const Body *a, const Body *b);
+float gravitational_force_magnitude(const Body *a, const Body *b);
 Vector2 gravitational_force(const Body *a, const Body *b);
 
 #endif
